Make kernel work with fingerprint HAL from B350
fingerprint.msm8916.so reads g_vendor_info_value from
"/sys/devices/platform/fingerprint/vendor_info".
When it is 0x30 (thats "0" value in file) it seems to does nearly the same
as previous HAL version. When it is 0x31 ("1" in file") it seems to
do additional initialization called "SYNA init".

For now hardcode 0 value.